Update structure of Application so things can be modular
