+++ spark/Helpers.py
@@ -8,10 +8,12 @@ from twilio.rest import Client
 current_path = os.path.abspath(os.path.dirname(__file__))
 
 config = ConfigParser.ConfigParser()
-config.read(os.path.join(current_path, '.env'))
+config.read(os.path.join(current_path, '../', '.env'))
 
 twilio_number = config.get('TWILIO', 'number')
 twilio_client = Client(config.get('TWILIO', 'sid'), config.get('TWILIO', 'auth'))
+
+env = config.get('APP', 'env')
 
 def send_message(number, message):
   return twilio_client.messages.create(
