+++ spark/templates/example_products.json
